<chore> Add .gitignore
Swap files (.swp / .swo) keep appearing due to closing the terminal without exiting vim.
This is because swap files are created to keep track of the changes made to a file before a crash.
It's probably good practise to still delete these swap files when they arrive, but now they won't show up in the unstaged changes.