<fix> Patch display putting + instead of * when converting expression to MathJax
Hard one to explain. Basically, this code needs to be refactored.
But, for now, there's a problem where some expression might satisfy the outer if statement to be treated as implied *, but not the inner. In this case, since the outer if was entered, the program flow cascades through the rest of the else-ifs, and down into the + section, where it would add a +.
Changing the else if's to just ifs fixes this.