<fix> Fix commutativity for nested commutative terms
Nested commutative terms: A commutative operator (and its children), that is itself a node that's commutative to some other operator above it in the tree.
e.g: In 1+2+3*4, * is a nested commutative operator on +.
The issue here is to do with omitting the binding declaration (let/const) in the for..of loop.
If no binding declaration is set, then js walks up the scopes (local -> global) to find if there's an exisiting variable with that name. If it doesn't find one, it creates a global variable of that name.
This explains what's happening here - the findCommutativeNodes function is called, and the for..of loop is hit. node doesn't exist, so it's created globally.
When the function recurs, and the for..of loop is hit again, the global variable node already exists. Therefore, its value is *overwritten*, which is not what we want.
We want this variable to be local to the loop, so use const or let!